KAN-64 Migrar Pages Router para App Router
