--- pages/recuperar/index.tsx
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-import Base from '@components/base-layout/index';
-import NewMenu from '@components/newMenu/index'
-import ResetPWD from '@components/reset'
-import FooterMain from '@components/footer/main/index';
-
-export default function Index() {
-	return (
-		<>
-			<Head>
-				<title>Recuperar Senha | GestBucal</title>
-			</Head>
-			<Base
-				appBarChild={<NewMenu/>}
-				mainContainerChild={<ResetPWD/>}
-				footerChild={<FooterMain />}
-			/>
-		</>
-	);
-}
