--- pages/team/index.tsx
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-import Base from '@components/base-layout/index';
-import NewMeni from '@components/newMenu/index'
-import AboutUsContainer from '@components/container/about-us';
-import FooterMain from '@components/footer/main/index';
-
-export default function Index() {
-	return (
-		<>
-			<Head>
-				<title>Equipe | GestBucal</title>
-			</Head>
-			<Base
-				appBarChild={<NewMeni/>}
-				mainContainerChild={<AboutUsContainer/>}
-				footerChild={<FooterMain />}
-			/>
-		</>
-	);
-}
