+++ src/app/articles/page.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { useEffect } from 'react';
+import Base from '@components/base-layout/index';
+import NewMenu from '@components/newMenu/index';
+import CollectionContainer from '@components/container/collection';
+import FooterMain from '@components/footer/main/index';
+
+export default function Page() {
+    useEffect(() => {
+        document.title = 'Artigos | GestBucal';
+    }, []);
+
+    return (
+        <Base appBarChild={<NewMenu />} mainContainerChild={<CollectionContainer />} footerChild={<FooterMain />} />
+    );
+}
