+++ src/app/faq/page.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { useEffect } from 'react';
+import Base from '@components/base-layout/index';
+import NewMenu from '@components/newMenu/index';
+import Faq from '@components/container/faq';
+import FooterMain from '@components/footer/main/index';
+
+export default function Page() {
+    useEffect(() => {
+        document.title = 'FAQ | GestBucal';
+    }, []);
+
+    return (
+        <Base appBarChild={<NewMenu />} mainContainerChild={<Faq />} footerChild={<FooterMain />} />
+    );
+}
