+++ src/app/planeja/page.tsx
@@ -1,0 +1,16 @@
+/* eslint-disable react/no-unescaped-entities */
+'use client';
+
+import React, { useEffect, useState } from 'react';
+import PlanForm from '@components/planeja/planeja-form';
+import { FinishedForm } from '@components/planeja/finished-form';
+
+export default function Page() {
+    const [isCompleted, setIsCompleted] = useState(false);
+
+    useEffect(() => {
+        document.title = 'Planeja | GestBucal';
+    }, []);
+
+    return isCompleted ? <FinishedForm /> : <PlanForm onFinish={() => setIsCompleted(true)} />;
+}
